CWL-to-JSON employed
We can now convert between CWL and JSON. While the cwlrunner implemented its own validation, now we can convert between JSON and CWL so we can tell what a front-end GUI can expect to see before running things. Additionally, we will not have to do the conversion on the front-end, but now we control it here.