Required vs Optional Inputs
Now that Workflows can be parsed into valid JSONSchemas. Users will be
prompted to to provide values for required inputs and be given the
option to pass on optional inputs. In addition, job templates are generated
not by complex loops but by the cwltool itself, which gives us the ability
to see whether the inputs are declared required or optional by the original
workflow file.